Add Docker config
I'm planning to deploy this through docker for my own convenience, so adding
docker config. I've updated the readme to reflect this, and to be explicit
about the option of using an environment variable for the token, which will
be necessary for this (and safer in general), and which the code already supports.